requirements.txt
pyyaml==6.0.1
numpy==1.26.0
pandas==2.1.1
openpyxl==3.1.2